RPi/requirements.txt
Adafruit-Blinka==3.7.0
adafruit-circuitpython-74hc595==1.0.3
adafruit-circuitpython-busdevice==4.1.1
adafruit-circuitpython-charlcd==3.2.2
adafruit-circuitpython-mcp230xx==2.2.1
Adafruit-PlatformDetect==2.1.0
Adafruit-PureIO==1.0.4
aniso8601==8.0.0
asn1crypto==0.24.0
certifi==2018.8.24
chardet==3.0.4
Click==7.0
cryptography==2.6.1
entrypoints==0.3
Flask==1.1.1
Flask-RESTful==0.3.7
idna==2.6
itsdangerous==1.1.0
Jinja2==2.11.0
keyring==17.1.1
keyrings.alt==3.1.1
luma.core==1.12.0
luma.led-matrix==1.4.1
MarkupSafe==1.1.1
numpy==1.16.2
opencv-python==3.4.6.27
Pillow==7.0.0
pycrypto==2.6.1
pyftdi==0.42.2
PyGObject==3.30.4
pyserial==3.4
python-apt==1.8.4
pytz==2019.3
pyusb==1.0.2
pyxdg==0.25
requests==2.21.0
rpi-ws281x==4.2.3
RPi.GPIO==0.7.0
SecretStorage==2.3.1
six==1.12.0
smbus2==0.3.0
spidev==3.4
ssh-import-id==5.7
sysv-ipc==1.0.1
urllib3==1.24.1
Werkzeug==0.16.1
